Separate out the configuration code from initialization
